backend/python/stream-processor/pyproject.toml
[project]
name = "stream-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "faster-whisper>=1.2.1",
    "grpcio>=1.76.0",
    "grpcio-testing>=1.76.0",
    "grpcio-tools>=1.76.0",
    "numpy>=2.4.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "uvicorn>=0.40.0",
    "redis>=5.0.1",
    "opentelemetry-distro",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-grpc",
    "opentelemetry-instrumentation-redis",
    "opentelemetry-instrumentation-logging",
    "av>=14.0.1",
]

[tool.setuptools]
py-modules = ["main", "transcriber", "cc_pb2", "cc_pb2_grpc"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
